requirements.txt
aniso8601==7.0.0
asgiref==3.3.4
beautifulsoup4==4.9.3
bovespa==0.1.0
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
dj-database-url==0.5.0
Django==3.2.3
django-configurations==2.2
django-filter==2.4.0
django-simple-history==3.0.0
djangorestframework==3.12.4
fundamentos==1.4
google-api-core==1.28.0
google-api-python-client==2.6.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gspread==3.7.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
inflection==0.5.1
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.2
more-itertools==8.8.0
multitasking==0.0.9
numpy==1.20.3
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
pandas==1.2.4
pandas-datareader==0.9.0
Pillow==8.2.0
promise==2.3
protobuf==3.17.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
Quandl==3.6.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.5
whitenoise==5.2.0
xlrd==2.0.1
yfinance==0.1.59
